Move HostConfiguration composition local into Redwood
It is not hooked up yet, but at least it is there.
