+++ redwood-treehouse-guest-compose/src/commonMain/kotlin/app/cash/redwood/treehouse/composition.kt
@@ -18,6 +18,7 @@ package app.cash.redwood.treehouse
 import androidx.compose.runtime.CompositionLocalProvider
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
+import app.cash.redwood.compose.LocalHostConfiguration
 import app.cash.redwood.compose.RedwoodComposition
 import app.cash.redwood.ui.HostConfiguration
 import kotlinx.coroutines.flow.StateFlow
